+++ torchao/utils.py
@@ -6,12 +6,12 @@ def benchmark_model(model, num_runs, input_tensor):
     start_event = torch.cuda.Event(enable_timing=True)
     end_event = torch.cuda.Event(enable_timing=True)
     start_event.record()
-    
+
     # benchmark
     for _ in range(num_runs):
         with torch.autograd.profiler.record_function("timed region"):
             model(input_tensor)
-    
+
     end_event.record()
     torch.cuda.synchronize()
     return start_event.elapsed_time(end_event) / num_runs
@@ -24,3 +24,18 @@ def profiler_runner(path, fn, *args, **kwargs):
         result = fn(*args, **kwargs)
     prof.export_chrome_trace(path)
     return result
+
+def get_compute_capability():
+    if torch.cuda.is_available():
+        capability = torch.cuda.get_device_capability()
+        return float(f"{capability[0]}.{capability[1]}")
+    return 0.0
+
+def skip_if_compute_capability_less_than(min_capability):
+    def decorator(test_func):
+        def wrapper(*args, **kwargs):
+            if get_compute_capability() < min_capability:
+                raise unittest.SkipTest(f"Compute capability is less than {min_capability}")
+            return test_func(*args, **kwargs)
+        return wrapper
+    return decorator
